Added Browser command menu
IPS starts in dual networking mode - Soft Access Point and Station mode. I use my smartphone to connect to it, it assigns an IP address, I use my browser to access an HTML-based command menu.

File uploads and deletes are supported using the browser. It shows what is on the SPIFFS file system.

Use the browser interface to select from a Wifi network (SSID) and connect. This makes it super easy (barely an inconvenience) to connect to a network to enable the NTP protocol to set the clock.